requirements-dev.txt
black
django-stubs
poetry
pytest
pytest-django
pytz
